feat(docker-tests): implement module gating with feature flags
Add comprehensive feature flag system for selective docker test compilation:

Feature flags added to Cargo.toml:
- docker-tests-qrc20: QRC20 coin tests
- docker-tests-tendermint: Tendermint/IBC coin tests
- docker-tests-zcoin: ZCoin/Zombie coin tests
- docker-tests-swaps-utxo: UTXO swap protocol tests (v1, v2, confs, locks)
- docker-tests-watchers: Watcher node tests (ETH + UTXO)
- docker-tests-ordermatch: Orderbook and matching tests

Test organization:
- Gate all test modules by category (ordermatching, swaps, watchers, coin-specific)
- Add detailed comments documenting test purposes and chain dependencies
- swap_tests runs only in main docker job via exclusion logic

Infrastructure cleanup:
- Move QtumDockerOps from qrc20_tests.rs to helpers/qrc20.rs
- Gate helper modules on run-docker-tests (env/eth also for sepolia)

All feature combinations verified to compile successfully.
